autocomplete working

+++ src/cursor_animation.rs
@@ -1,8 +1,8 @@
 use crate::events;
 
 pub struct CursorAnimation {
-    target_pos: (u16, u16),
-    prev_target_pos: (u16, u16),
+    target_pos: (usize, usize),
+    prev_target_pos: (usize, usize),
     tick_of_change: u64,
 }
 
@@ -15,16 +15,15 @@ impl CursorAnimation {
         }
     }
 
-    pub fn update_position(&mut self, new_pos: (u16, u16), tick: u64) {
+    pub fn update_position(&mut self, new_pos: (usize, usize), tick: u64) {
         if new_pos != self.target_pos {
             self.tick_of_change = tick;
             self.prev_target_pos = self.target_pos;
-            // self.Prev_target_pos = self.get_position(tick);
             self.target_pos = new_pos;
         }
     }
 
-    pub fn get_position(&self, tick: u64) -> (u16, u16) {
+    pub fn get_position(&self, tick: u64) -> (usize, usize) {
         // interpolate between prevPos and currentPos based on time since tick_of_change
         let ticks_since_change = tick.saturating_sub(self.tick_of_change);
         let mut factor =
@@ -39,7 +38,7 @@ impl CursorAnimation {
             + (self.target_pos.0 as f32 - self.prev_target_pos.0 as f32) * factor;
         let y = self.prev_target_pos.1 as f32
             + (self.target_pos.1 as f32 - self.prev_target_pos.1 as f32) * factor;
-        (x as u16, y as u16)
+        (x as usize, y as usize)
     }
 
     pub fn get_intensity(&self, tick: u64) -> u8 {
@@ -47,6 +46,5 @@ impl CursorAnimation {
         let intensity_f32 = (tick as f32 * mult).sin() * 0.4 + 0.6;
         let intensity = (intensity_f32 * 255.0) as u8;
         intensity
-        // 255
     }
 }
